Add the interface for parameter dependent functions
This introduces a new class pdouble, which represent
parameters. These are supposed to work in such a way
that they are really constants, which can be changed
from one evaluation of the function to the next, but
we do not compute derivatives with respect to these
parameters. This will speed up forward mode significantly
as it will reduce the number of real independent variables.

We also move the declarations of overloaded functions
to a separate file because these friend declarations
were repeated between badouble and adub classes, and now
new such functions with pdouble parameters are added.
In order to keep the declarations in sync we include
the extra internal headers inside each class declaration.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>